+++ node/src/lib.rs
@@ -1,6 +1,7 @@
 #[macro_use]
 extern crate diesel;
 
+pub mod chain;
 pub mod config;
 pub mod opt;
 pub mod store_builder;
